Refactor into new Data class
The new class is meant to encapsulate the data and it's operations from
the application. Reading, writing, modification, and removal should be
implemented here in future revisions. The major difference between
classes is Pandas instead of the csv module implementation I previously
used. This should give us many more options down the road.