Extend Data functionality
The read_data method has been expanded to accept *args, and **kwargs to
pass along to pandas.read_csv(). This will allow further flexibility in
how data is read from disk.

Add a basic write_data method. This simply wraps and passes arguments
along to pandas.to_csv(). Testing revealed an issue where the data
written back to disk is in the same format as it is retrieved in. This
causes the date and time columns to be merged into a single 'burned at'
column. Originally this was not desired behavior. I am wondering now if
it may apply for 'Feature -- not a bug' status.