Update CSV format, again.
Rationale behind this change is to simplify the format for use with
Pandas. Merging the two columns together should allow sorting by
date/time to be a bit easier.

In all likelyhood though, this format will not be solidified until after
we implement the reading/writing of data files. We will almost certainly
stumble upon a better format for our data.