Simplify calls to read_data and write_data
By removing some of the extra arguments passed to read_data, we
eliminate the custom formatting applied to our CSV data. This will be
useful until we are able to create a fix for Issue 1, linked below.

Uncommenting the call to write_data will be useful to check our output
is the same as our input in cases where it should not be modified. Also
useful in eliminating Issue 1.

https://github.com/ncdulo/pymaengda/issues/1