Fix Issue 1 by specifying column to use as index
Passing the index_col=0 argument to read_data appears to keep our data
in the same format for both reading and writing. The output files show
no difference between the two and I feel mostly confident the issue is
fixed. However, changing the CSV format, formatting the DataFrame or
otherwise transforming the data will likely result in non-original data
to be written back to disk.

https://github.com/ncdulo/pymaengda/issues/1