Add more skeleton code
Instead of displaying debug text, specifying a runlevel now calls the
appropriate function. These functions will implement the core
functionality of our program. I am still undecided on this format. I may
try an alternative format with Application.run() handling data
read/write from disk instead of the individual functions. That is
probably not a very pythonic methodology though.

Moved data_dir, data_file up in scope to be Application class members.
The rationale behind this is that the values are accessed multiple
times, and the value is constant between instances of the application.