Create read_data, write_data
These two functions serve to read or write the data file. I decided to
make them into individual functions to save on repetition of code.
Because the possibility exists where the program mode of operation may
require different arguments to the underlying Pandas calls, we are
adding *args, **kwargs to this top-level wrapper.