Change read_data, write_data arguments.
Here I have changed the read_data, and write_data function calls to
accept an argument for infile, and outfile, respectively. This will
allow us to specify specific files to read or write from, depending on
the situation being handled.

In addition to the function changes, we now determine the data_file used
by those calls once. The original behavior would re-compute those
strings in each runlevel. Again this will allow for more flexibility,
and leave less room for error in the future.