Modify write_data to use internal copy of data
Data.write_data will now utilize the copy of the data stored within the
Data class instead of relying on Application to pass it back in. The
reasoning behind this is that there is only one set of data. We do not
have a need to modify this data outside of Data class. Various
modifications such as add row, delete, modify, etc will be handled
within Data. As these actions may be specific to the data type or file
format, they should be separated from the main Application logic.

This change also facilitates a touch of optimization by not passing
around un-needed copies of the data as much.