Data now stores it's own file names
Added in read_file, write_file to the Data class to hold the file to
the files for reading and writing to. If the write file is not specified
during object instantiation, it will default to being the same as
read_file.

Following the logic that we have only one set of data, that data is
contained within specific files. By storing the file names alongside the
data we cut down on argument passing and tighten up the allowed room for
error.

Making this change also allows us to remove the member variables used in
Application to determine and store those file names.