Begin implementing argument sub parsers
Using the sub-parsers provided by argparse we are able to achieve nicer
argument formatting. The basics of it is that users will select a mode
for the program to operate in followed by any required arguments. This
is done without using flags for the required arguments, only for the
optional ones.

This enables 'nicer' commands such as:
  sh bin/pymaengda burn 3.5 super-green
or
  sh bin/pymaengda log --limit 25