Add base 'edit' mode.
Currently not implemented as I have yet to determine how this mode
should work. Will we have ID numbers stored with the data for easy
selection?

That seems to be a good way for handling things. Would require an update
of the CSV data format to include a proper unique ID but that is very
doable.

We likely will not support any editing of timestamps at this point. The
timestamp should be seen as a constant with each entry. We should be
able to optionally specify it when creating an entry, but modifying it
after the fact I'm unsure on. Mostly because it potentially could place
the entries out of order with ID numbers not lining up with the
timestamp. It could be implemented, but may require some data-wrangling
to do so properly.