Sub-parser 'stats' created
The 'stats' sub-parser is used to select statistics mode. This will
allow display of various statistics and information on the CSV data. For
now, only one optional flag '--full' exists to specify a complete output
over the standard output. For now, neither is implemented. This is only
the mode selection for now.