Remove 'runlevel' argument/loop
It seems un-necessary to use the runlevel system and while loop to
control program flow in this case. In general, the program will only
loop once and be done. The cases where we wish to display the data file
after our action can be called manually. And removing the while loop
removes a chance for an infinite loop if something goes wrong.

Finally, being that we implemented a set of sub-parsers to select the
mode, we no longer have a need for a separate runlevel argument. This
can be thought of as an extension and piece of the sub-parsers
implementation.