Modify member assignment in Data.__init__
In the name of greater readability, change `self.write_file` to be
assigned the default value from `self.read_file`. This is a slightly
more verbose assignment than the previous one. That'll help me figure
out what this code is doing when I come back to it in the future. How's
this for verbosity?