Fix wrong file/dir names in Installation
After updating the project name as part of Issue #12, I intended to
update the installation instructions to reflect the new project
structure. After being caught up so long with mis-guided directories, I
forgot to do so. This commit fixes that. Including as part of PR #16 as
it falls into the same direction.