Default to list mode when no arguments given
This commit resolves Issue #10. When Suppylement is run with no command
line arguments provided, we will enter `list` mode by default. This is a
bit of an easy alias for a quick glance at the most recent entries.

To implement this feature, we check the length of the `sys.argv` list.
If there are less than two items (program name, argument, [argument,
..]`, we call `parse_args()` with our own argument string.