Resolve issue #7
This commit introduces the ability to pass custom argument lists into
`Application` instantiation. The main intention of this is to allow more
flexibility with unit testing. Which we still need to implement.

To accomplish this, we make several checks on the `args` passed in
during object creation. If we are given a custom argument string, use
that. Otherwise if we have been given `None`, default to command line
args. If that list does not include any actual arguments, we default to
`list` mode as a fallback.

This commit also undoes the changes made in PR #17, as I figured may
happen. `Application` seems like a better home for our argument checking
logic as that will be more specific to the application. `Arguments` is
more intended for the actual parser specification and calling.