Pep8 compliance check!
Remove 5 newlines, add 1 newline. Pep8 compliance check recommended
those changes. Glad to see there weren't more changes to be made.

Note I am ignoring the line too long error in `arguments.py` as that is
a line of debug text that will not be in future versions. And mostly
because I can't think of the right way to fix it this early in the
morning.