Add indices to CSV data
As I begin work on `Application.create()` and the underlying function
within `Data.new_entry()`, I am thinking again that we should include a
numeric index for each entry. It allows for an easier implementation of
`new_entry()` and I feel it will allow for easier working with the data
in the future.