Create Pandas DataFrame for new entries
The functionality introduced in this commit partially implements the
`Application.create()` method as part of Issue #20. Much of the actual
grunt work is handled in `Data.new_entry()` once given the arguments
from `Application.create()`.

This code will accept an `amount`, and `name` variable and convert into
a new Pandas DataFrame. We then append that new DataFrame onto our
existing `self._data`, which was previously read from disk.

Debug text is remaining in for now as work is continuing. This commit
merely represents a bit of a progress milestone.