Finalize core `new_entry` functionality
With the current timestamp being correctly added into new entries, and
the new entry properly being appended on to `self._data`, we can call
this method complete.

To generate the timestamp we call `datetime.now()` with a custom format.
This may be a good candidate for a value kept in the configuration file
when we get to implementing that.

Otherwise, the code is mostly the same as the previous commit. The main
difference is that we actually set `self._data` and we have changed the
text output slightly.

There is also a fix in `Data.__init__()` where if not given a
`write_file` argument, setting it would fail.