Add blank.csv
This file represents an empty data file, ready for use by Suppylement. I
am including as a separate file so that users may simply copy this over
to `data.csv` to begin using the program without worry as that file is
listed in .gitignore. Or rather will be within the new commit or two.