Check file(s) exist during `Data.__init__`
Check for the existence of our CSV file to read early on. Early to error
out, less likely for other problems to occur. The `Data.write_file`
member is *not* checked because it will be created if necessary when we
go to write data back to disk.

Otherwise this is just a simple 'file exists?'-type check that will
throw an Exception if the file is not found.