Capture exceptions in `read_data`
Exceptions and errors now handled in `Data.read_data` along with a check
in `Application.display` for an error condition. The
`pandas.errors.ParserError` and `pandas.errors.EmptyDataError`
exceptions may be thrown by `pandas.read_csv` to indicate errors
reading the data. We now properly capture those exceptions and exit
gracefully when encountered.

Finally, we add a check for `None` being returned by `Data.read_data`
and if so, display an error message.