Add basic argument support to list mode
This implements a basic framework for handling the data, and filtering
based on command line argument input, then finally displaying that data.
Instead of our old behavior to display all data entries, we filter down
the data.

First off, checking to see if we were called either directly into `list`
mode, or if we were called from another mode on the command line. We do
this because if we were originally called in a differnt mode, there will
be a different set of arguments read into Suppylement. In the case of
being called like that, we fall back to a default of 'most recent five
entries'.

Otherwise, we have reached this point by directly calling `list` mode. A
quick sanity-check of `self.args.most_recent` to ensure it is valid,
then we grab the proper rows of data after transforming the argument
value.

Regardless of which path we took above, we close off the function by
displaying the data or an error message if we have no data.