`most_recent` set to -1 shows all entries
Add an accepted value for `most_recent` to signal display of all entries
from the data file. `-1` was chosen because it will not be used under
normal circumstances, and fits in with our current argument parsing and
input setup. This will be useful during searches when all results may be
wanted for display.