Implement search by name
List mode now has the ability to search for entries matching a specified
name on the command line. The search is an exact search at the moment. I
would like to expand that to a fuzzy search that would be along the
lines of `*SEARCH_TERM*`.

To implement this feature, we add a new command line argument `--name`
with a default of `''`. If specified, we filter our data by that name
before entering our limit/most-recent filtering. We filter before the
most-recent slice otherwise we would only be working with a limited set
of data -- not intended behavior.

Another change worth mentioning is the shift from using `display_data`
to working entirely on `data`. This does over-write the original data
read from disk. However this is a local copy of the variable and will
not effect the data on disk unless `Data.write_data` is called without
re-reading the data. The intention behind this change is to ease the
further implementation of features.