Error check for amount filters. Add documentation.
Add simple error check along the lines of how we handle
`self.args.most_recent` error checking. If the `--less`, or `--more`
arguments are negative or zero, raise a ValueError alerting to the
condition. This will prevent unspecified behavior from entering
negative, or strangely ordered limits.

The search is also slightly updated to be inclusive for `--more` and
exclusive for `--less`.