Pep8 compliance check! Plus comments!
Took a glance over the code and added in some commentary to describe
some sections a bit better. Finding the right balance of code to
comment, and knowing when a comment is not needed is tricky sometimes.
Practice makes perfect!

Also hit the code with some `pep8` action and fixed a couple long line
warnings I was getting hit with.