Implement delete row functionality
Bit of a larger commit than I try to go for. Got a little carried away
as I had a few "Oh shoot! It actually works!" moments.

A long awaited feature, `rm` mode, has been created in it's most basic
form. You may pass in an `--id` argument on the command line and that
row will be deleted from disk. There are no "yes or no" confirmations,
choose wisely. An interactive/confirm flag is not a bad idea though.

Otherwise, we do some basic bounds checking to ensure the index is
within the correct range. Drop the row. Re-flow the index so all of it's
values are continuous. This re-flow prevents gaps or jumps in the index
numbers. An unintended side effect of this is a bit ease-of-use. When
deleting multiple rows, if you start towards lower numbers (ID 20
instead of 25), you can continue entering the same command without
changing the ID to remove consecutive rows. The ID should be confirmed
in the 'tail' output of the data after each removal.

Long awaited feature, simple implementation. See how easy that is once
you get down and do it?