Confirm delete flag argument
An optional flag argument added, `-i`, indicating an interactive `rm`
mode should be called. This interactive mode will select the `DataFrame`
and display it to the user, asking for confirmation. We are a little bit
picky about the input we accept here in an attempt to guard against
un-wanted data loss.

Success of the deletion is indicated by a True return value from
`Data.delete_row_by_id`, which is checked for and an appropriate message
taken. In this situation, if we succeed -- write the data to disk. If
the confirmation fails, inform the user but write no data.