Touch blank config file
Add `suppylement_defaults.ini` as a container to hold all of
Suppylement's default values. This file is not intended to be edited by
the end-user, it is a source of fall-back values for the program and a
basis for creating an individual user config.

User configuration should be specified in `suppylement.ini`. Any values
specified there will over-write values specified here in
`suppylement_defaults.ini`.