Add `Configuration` with empty `configparser`
A new class to manage the application configuration created,
`Configuration`. This class will handle the reading of multiple
configuration files and merging the values together with user specified
values taking precedence.

The current implementation is bare-bones and not functional for any true
purpose. It does, however, determine the proper file paths and then
reads those file. Creating a `configparser` object as it does so. This
is a very rudimentary setup and needs much work. But it's a start.