Add new values to default config
A few changes to the original default config have taken place. Just
growing this out as we need.  The original `data` section has been
renamed to `filenames` as I feel it more accurately describes the data.
`Data` felt a bit too broad for my liking here. A `defaults` section has
also been added with two unused values to start. The `custom_args`
provides a default list of arguments to pass in to `Application` if none
are specified on the command line. `confirm_delete` will modify the
default behavior of `rm` mode to always prompt before removing a row.

Safety first!