Instantiate `Data` using config file values
As part of `Configuration` object creation, assemble the full
`read_file` and `write_file` paths, using them during `Data` object
creation. This allows Suppylement to determine it's data file using
either the default, or a user-specified configuration file.

This allows us to remove a small block of code at the beginning of
`Application.__init__` related to determining data paths. The value is
now read from `suppylement_defaults.ini` or `suppylement.ini` with the
latter taking priority.

Several lines of debug text remains in the bottom of
`Configuration.__init__`. That will be removed as we hammer out the
finer details.