Update default `custom_args`
The default set of command line arguments to run Suppylement with when
none are provided has been updated. This update removes the
`--most-recent` flag as it was only intended for testing. Testing shows
it works, thus the flag is being removed from the defaults. Back to our
intended behavior of defaulting to list mode with a five row display.