Utilize `custom_args` config value
When Suppylement is run without any command line arguments, use the
value read from the config file to specify our defaults. It accepts any
valid argument string as it would be input on the command line. The
intention of this feature is to provide the ability to set a default
behavior for Suppylement. Generally this will be limiting or filtering
the output of list or statistics mode in some way.

If the `custom_args` value does not exist in the user, or program
default config file for some reason, fall back to help mode as a last
resort.