`confirm_delete` option implemented
Very small, simple commit. I wish I could say this was quick to write
up, but I did spend some time thinking it through. Pulling up the `diff`
just now to write the commit message surprised me with how short it is
in comparison to time spent.

Ah well, it works. There exists within the default configuration, an
option called `confirm_delete`. That option is a boolean value. When
turned on, enabled, set to True, or otherwise given a high value, we
will confirm all data deletions with the user before performing them.
This option is additive with the `-i` command line flag. That is,
setting one *or* the other will result in confirmation before any delete
operation is carried out.