Expanded comment block
Realizing that our 'data cache' never actually made the required check
to use it, we are also realizing we don't want that cache after all.
Gearing up for future development, we want the ability to re-read our
data and reset any transformations made to it. This commit only
introduces more documentation regarding the situation.