Prefer `pathlib` over `os` for handling file paths
This commit represents the bulk of the migration to using `pathlib`.
`Configuration` handles the bulk of the work resolving file paths and
directories and I chose a slightly larger commit as I did not want to
commit code that throws exceptions.

Changes in the configuration class are fairly straight-forward. Replace
the calls to `os` and update the code that builds our final filenames.
These filenames are used to read our configuration from disk, as well as
to read/write our actual supplement data. It took a bit of banging my
head, and documenation delving to get over some minor hangups. But it
appears to work now.

We also update the filenames on `application.py` and
`suppylement_defaults.py` by dropping any leading forward slashes. This
caught me up a bit before I realized that `pathlib` was interpreting the
leading forward slash as referring to the root directory -- not what we
want! The values have been updated and all is right in the world.