Move statement outside of block, update comments.
The initializing statement for `confirm_delete` is better located in the
top of `Application.delete` for readability. The function works as
intended regardless of whether this statement is located in the
`try/except` block as it was originally, or if it is moved up a level to
this new location. This is simply an update for clarity of code.

The comment block next to the statement has also been updated for
clarity as well. Previously it was describing some of our original
logic.

Finally, a comment in `Data.new_entry` had some verbage corrected as
'application will decide' is more correct than 'application much
describe'.