Barebones, nested dictionary iteration
Let's (finally!) start working to migrate our data filters over to the
`Data` class. First off, write a barebones `Data.filter_data` method
which will accept a nested dictionary of columns and filters to apply to
the column. Write a couple `for` loops, and verify we are properly
reading the data from the dictionary. Implement a test function (should
probably be a unit test somewhere...). Feel satisfied that it works in
the most basic sense. Sleep and think of how to proceed.

The thinking bit is to think through how we actually handle the filters
within our innermost loop. The easiest, would probably be a series of
`if` statements checking the filter type (`if filter_key ==
'less_than'...`). I'm not sure how efficient this feels. We can probably
do better. And we can probably go more generic.

For now, it's late and the brain is ceasing to work at the moment. For
now, good work. Getting started was the big hurdle. Build that momentum
back up!