Migrate from wrapper script to `setuptools`
Bringing in one main feature, with several minor changes along with it.
This commit introduces a functional `setup.py` which can be used to
install Suppylement. Very basic for now. Just enough that it works.
Sooner or later, I will flesh this out with the proper metadata.

Using the entry-points feature of `setuptools`, we no longer need to use
the script located in `bin/suppylement`. An executable will be created
when installing via Pip.

For now, I am leaving the shell script in the repo. Some users may
prefer it. In line with that note, I slightly modified our `main.py` to
include a proper entry-point function. Also leaving in the `__name__ ==
__main__` idiom for any user who wishes to run the `main.py` directly.

May want to consider using `__main__.py` as well for `python -m
suppylement` use. Too many options? Not enough options? I don't know.
It works for me. Pull requests, or issues are welcome if it doesn't work
for you.